Move FormData into form API
