+++ extension/ts/util/bookForm.ts
@@ -1,4 +1,9 @@
+type FormData = Record<string, Record<string, any>>;
+
 const FORM_DATA_ELEMENT_TAGS = ["textarea", "input", "select"];
+
+const COLLECTIONS_ID_PREFIX = "collection_u_";
+const COLLECTIONS_KEY = "___collections_";
 
 const getElementsByTag = (parent: HTMLElement) => (tag: string) => Array.from(parent?.getElementsByTagName(tag) ?? []);
 
@@ -13,4 +18,49 @@ const formExists = (): boolean => !!document.querySelector("#book_editForm > .bo
 
 const getForm = (): HTMLElement => document.getElementById("book_editForm");
 
-export {getForm, getFormElements, formExists};
+const getFormData = () => getFormElements()
+	.reduce((saveData: FormData, element: any) => {
+		// We can't change hidden elements because LibraryThing relies
+		// on hidden form inputs to send additional, form-specific metadata
+		// on save
+		if (element && element.id && element.type !== "hidden") {
+			const {value, checked} = element;
+			if (element.id.startsWith(COLLECTIONS_ID_PREFIX)) {
+				const collections = saveData[COLLECTIONS_KEY] || {};
+				const [span] = element.parentElement.getElementsByTagName("span");
+				collections[span.textContent] = {value, checked};
+				saveData[COLLECTIONS_KEY] = collections;
+			} else {
+				saveData[element.id] = {value, checked};
+			}
+		}
+		return saveData;
+	}, {});
+
+
+const extractSaveDataFor = (targetElement: Element, saveData: FormData) => {
+	if (targetElement.id.startsWith(COLLECTIONS_ID_PREFIX)) {
+		const span = targetElement.parentElement.getElementsByTagName("span")[0];
+		return saveData[COLLECTIONS_KEY][span?.textContent] ?? targetElement;
+	} else {
+		return saveData[targetElement.id] ?? targetElement;
+	}
+};
+
+const insertFormData = (saveData: FormData) => getFormElements()
+	.forEach((element: any) => {
+		// We can't change hidden elements because LibraryThing relies
+		// on hidden form inputs to send additional, form-specific metadata
+		// on save
+		if (element && element.id && element.type !== "hidden") {
+			const {value, checked} = extractSaveDataFor(element, saveData);
+			if (element.value !== value || element.checked !== checked) {
+				element.value = value;
+				element.checked = checked;
+				element.dispatchEvent(new Event("change"));
+			}
+		}
+	});
+
+export type {FormData};
+export {getForm, getFormElements, formExists, getFormData, insertFormData};
